refactor(app): extract global selection expansion context menu wiring
